Only set base time if hardware time is positive
This avoids the case where the initial hardware time is a very small but
negative value, which causes the offset to be equal to the initial
ros::Time::now(), which results in a zero shift in the compensated
hardware timestamps. This check forces SyncBase to wait until hardware
time is positive before setting the offset, which results in a nonzero
shift in the compensated hardware timestamps.