Fix bug in previous commit
If tune_exposure_time was set to a negative value, the driver attempted
to wait a negative amount of time before clamping the exposure time to
a fixed value. This resulted in auto-exposure being enabled permanently.
This change only starts the auto exposure time interval when
tune_exposure_time is set to a positive value.